Add files via upload
Fixes to *-list-deallocate functions.
Development of region-list-corr region lists.  Where the regions in the list correspond to the domains in the session domain list.
The goal is that all domains can be considered when changing all domain current states.
Positive and negative RLCs can be defined.
Negative RLCs can be avoided, as much as possible, when changing all current states.
When all current states are in a negative RLC, a need can be made to exit it.
When all current states are not in a positive RLC, a need can be made to seek one.
When all current states are in a positive RLC too long (boredom/satiation), another positive RLC can be sought.
Also develop functions for lists of RLCs.
RLCs add significant complication, but provide a basis for behavior.