Add files via upload
Continued development.  Added struct.fs and structlist.fs, as part of the Memory Management project.  For planning a path between region-list-corr instances, to remain within a rate negative quantity limit, using intersections, is very close to finding a rule-path as already coded.  A rule from a region-list-corr instance to an intersection within same can be expressed as a rule, with a few X->0 and X->1 changes.